backend/requirements.txt
boto3==1.28.68
fastapi==0.104.0
httpx==0.25.0
moto==4.2.6
pydantic==2.4.2
pytest==7.4.2
pytest-env==1.0.1
pytest-cov==4.1.0
python-dotenv==1.0.0
requests==2.31.0
requests-mock==1.11.0
six==1.16.0
sniffio==1.3.0
starlette==0.27.0
tomli==2.0.1
typing_extensions==4.8.0
urllib3==2.0.6
uvicorn==0.23.2
spacy==3.7.0
mysql-connector-python==8.2.0